requirements.txt
google-cloud-bigquery
python-dotenv
pillow
discord.py
requests